+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/async_samples/sample_copy_model_async.py
@@ -66,7 +66,8 @@ class CopyModelSampleAsync(object):
 
         async with target_client:
             copied_over_model = await target_client.get_custom_model(copy.model_id)
-            print(copied_over_model)
+            print("Model ID: {}".format(copied_over_model.model_id))
+            print("Status: {}".format(copied_over_model.status))
         # [END copy_model_async]